+++ plugins/scaffolder-backend/src/scaffolder/tasks/DatabaseTaskStore.ts
@@ -15,8 +15,10 @@
  */
 
 import { JsonObject } from '@backstage/types';
-import { PluginDatabaseManager } from '@backstage/backend-common';
-import { resolvePackagePath } from '@backstage/backend-plugin-api';
+import {
+  DatabaseService,
+  resolvePackagePath,
+} from '@backstage/backend-plugin-api';
 import { ConflictError, NotFoundError } from '@backstage/errors';
 import { Knex } from 'knex';
 import { v4 as uuid } from 'uuid';
@@ -78,19 +80,19 @@ export type RawDbTaskEventRow = {
  * @public
  */
 export type DatabaseTaskStoreOptions = {
-  database: PluginDatabaseManager | Knex;
+  database: DatabaseService | Knex;
   events?: EventsService;
 };
 
 /**
- * Type guard to help DatabaseTaskStore understand when database is PluginDatabaseManager vs. when database is a Knex instance.
+ * Type guard to help DatabaseTaskStore understand when database is DatabaseService vs. when database is a Knex instance.
  *
  * * @public
  */
-function isPluginDatabaseManager(
-  opt: PluginDatabaseManager | Knex,
-): opt is PluginDatabaseManager {
-  return (opt as PluginDatabaseManager).getClient !== undefined;
+function isDatabaseService(
+  opt: DatabaseService | Knex,
+): opt is DatabaseService {
+  return (opt as DatabaseService).getClient !== undefined;
 }
 
 const parseSqlDateToIsoString = <T>(input: T): T | string => {
@@ -152,9 +154,9 @@ export class DatabaseTaskStore implements TaskStore {
   }
 
   private static async getClient(
-    database: PluginDatabaseManager | Knex,
+    database: DatabaseService | Knex,
   ): Promise<Knex> {
-    if (isPluginDatabaseManager(database)) {
+    if (isDatabaseService(database)) {
       return database.getClient();
     }
 
@@ -162,10 +164,10 @@ export class DatabaseTaskStore implements TaskStore {
   }
 
   private static async runMigrations(
-    database: PluginDatabaseManager | Knex,
+    database: DatabaseService | Knex,
     client: Knex,
   ): Promise<void> {
-    if (!isPluginDatabaseManager(database)) {
+    if (!isDatabaseService(database)) {
       await client.migrate.latest({
         directory: migrationsDir,
       });