+++ plugins/app-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-app-backend
 
+## 0.4.4-next.0
+
+### Patch Changes
+
+- d9d62ef: Remove some internal usages of the backend-common package
+- 8379bf4: Remove usages of `PluginDatabaseManager` and `PluginEndpointDiscovery` and replace with their equivalent service types
+- Updated dependencies
+  - @backstage/config-loader@1.9.5-next.0
+  - @backstage/plugin-auth-node@0.5.6-next.0
+  - @backstage/backend-plugin-api@1.1.1-next.0
+  - @backstage/config@1.3.1
+  - @backstage/errors@1.2.6
+  - @backstage/types@1.2.0
+  - @backstage/plugin-app-node@0.1.29-next.0
+
 ## 0.4.3
 
 ### Patch Changes