+++ plugins/events-backend-test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-events-backend-test-utils
 
+## 0.1.40-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-events-node@0.4.7-next.0
+
 ## 0.1.39
 
 ### Patch Changes