+++ plugins/events-backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-events-backend
 
+## 0.4.1-next.0
+
+### Patch Changes
+
+- d9d62ef: Remove some internal usages of the backend-common package
+- Updated dependencies
+  - @backstage/backend-openapi-utils@0.4.1-next.0
+  - @backstage/backend-plugin-api@1.1.1-next.0
+  - @backstage/config@1.3.1
+  - @backstage/errors@1.2.6
+  - @backstage/types@1.2.0
+  - @backstage/plugin-events-node@0.4.7-next.0
+
 ## 0.4.0
 
 ### Minor Changes